Revert "feat: support entity-bytes via daemon, add unixfs fixture itests"
This reverts commit 4e4a578ed7d44417521db59b92c3e46dc0ca647d.
