+++ pkg/storage/duplicateaddercar_test.go
@@ -19,6 +19,7 @@ import (
 )
 
 func TestDuplicateAdderCar(t *testing.T) {
+
 	setupStore := &testutil.CorrectedMemStore{ParentStore: &memstore.Store{
 		Bag: make(map[string][]byte),
 	}}
@@ -33,7 +34,7 @@ func TestDuplicateAdderCar(t *testing.T) {
 
 	store := storage.NewDeferredStorageCar("")
 	ctx := context.Background()
-	carWriter := storage.NewDuplicateAdderCarForStream(ctx, unixfsFileWithDups.Root, "", types.DagScopeAll, nil, store, buf)
+	carWriter := storage.NewDuplicateAdderCarForStream(ctx, unixfsFileWithDups.Root, "", types.DagScopeAll, store, buf)
 	cachingTempStore := storage.NewCachingTempStore(carWriter.BlockWriteOpener(), store)
 
 	// write the root block, containing sharding metadata
