+++ pkg/types/request.go
@@ -109,14 +109,7 @@ type RetrievalRequest struct {
 // and writing and it is explicitly set to be trusted (i.e. it will not
 // check CIDs match bytes). If the storage is not truested,
 // request.LinkSystem.TrustedStore should be set to false after this call.
-func NewRequestForPath(
-	store ipldstorage.WritableStorage,
-	cid cid.Cid,
-	path string,
-	dagScope DagScope,
-	byteRange *ByteRange,
-) (RetrievalRequest, error) {
-
+func NewRequestForPath(store ipldstorage.WritableStorage, cid cid.Cid, path string, dagScope DagScope) (RetrievalRequest, error) {
 	retrievalId, err := NewRetrievalID()
 	if err != nil {
 		return RetrievalRequest{}, err
@@ -135,7 +128,6 @@ func NewRequestForPath(
 		Cid:         cid,
 		Path:        path,
 		Scope:       dagScope,
-		Bytes:       byteRange,
 		LinkSystem:  linkSystem,
 	}, nil
 }
@@ -205,15 +197,11 @@ func (r RetrievalRequest) GetUrlPath() (string, error) {
 	if legacyScope == string(DagScopeEntity) {
 		legacyScope = "file"
 	}
-	byteRange := ""
-	if !r.Bytes.IsDefault() {
-		byteRange = "&entity-bytes=" + r.Bytes.String()
-	}
 	path := r.Path
 	if path != "" {
 		path = "/" + path
 	}
-	return fmt.Sprintf("%s?dag-scope=%s&car-scope=%s%s", path, scope, legacyScope, byteRange), nil
+	return fmt.Sprintf("%s?dag-scope=%s&car-scope=%s", path, scope, legacyScope), nil
 }
 
 // GetSupportedProtocols will safely return the supported protocols for a specific request.
